requirements.txt
openai
phidata
duckdb
yfinance
sqlalchemy
pgvector
psycopg2-binary
pypdf
python-dotenv
streamlit
pandas
